Create sub-packages inside duke package
