+++ src/main/java/duke/components/MainWindow.java
@@ -1,5 +1,7 @@
-package duke;
+package duke.components;
 
+import duke.Duke;
+import duke.ui.Ui;
 import javafx.animation.PauseTransition;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
@@ -25,7 +27,7 @@ public class MainWindow extends AnchorPane {
     private Button sendButton;
     private Ui ui = new Ui();
 
-    private Duke duke;
+    private Duke duke = new Duke();
 
     private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
     private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
