+++ src/test/java/duke/TodoTest.java
@@ -2,6 +2,8 @@ package duke;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import duke.exceptions.DukeException;
+import duke.tasks.ToDo;
 import org.junit.jupiter.api.Test;
 public class TodoTest {
     @Test
